Исправление ошибок работы с сервером
