+++ js/main.js
@@ -1,8 +1,15 @@
-import { importImageData } from './import-data.js';
-import { addUploadHandlers } from './image-download.js';
-import { postImage } from './image-posting.js';
+import { loadPhotos, addImageSubmitHandler } from './api.js';
+import { addUploadHandlers } from './download.js';
+import { renderImages } from './render.js';
 
-importImageData();
+const BASE_URL = 'https://22.javascript.pages.academy/kekstagram';
+const DESCRIPTIONS_COUNT = 25;
+
+loadPhotos(`${BASE_URL}/data`)
+  .then((descriptions) => {
+    renderImages(descriptions.slice(0, DESCRIPTIONS_COUNT));
+  });
+
 addUploadHandlers();
-postImage();
+addImageSubmitHandler(BASE_URL);
